The graphical field simulator home
This is going to be the home of the graphical field simulator. I feel
that it would be better organised if all the data was in another folder,
so here it is.
I have currently completed task 1, which involved making the graphics
field item class and the graphics crop item class.